@@ -1,6 +1,6 @@
 {
   "name": "ngx-translate-cache",
-  "version": "17.0.0",
+  "version": "18.0.0",
   "description": "ngx-translate extension to facilitate language cache.",
   "repository": {
     "type": "git",
@@ -24,8 +24,8 @@
     "url": "https://github.com/jgpacheco/ngx-translate-cache/issues"
   },
   "peerDependencies": {
-    "@angular/common": "^17.0.0",
-    "@angular/core": "^17.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/core": "^18.0.0",
     "@ngx-translate/core": ">=14.0.0"
   }
 }
