@@ -1,43 +1,38 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { TranslateModule, TranslateLoader, TranslateService } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
-import { TranslateCacheModule, TranslateCacheSettings, TranslateCacheService } from 'ngx-translate-cache';
+import { TranslateCacheModule, TranslateCacheSettings, TranslateCacheService } from 'ngx-translate-cache-2';
 
 import { AppComponent } from './app.component';
 
 export function HttpLoaderFactory(http: HttpClient) {
-  return new TranslateHttpLoader(http);
+    return new TranslateHttpLoader(http);
 }
 
 export function TranslateCacheFactory(translateService, translateCacheSettings) {
-  return new TranslateCacheService(translateService, translateCacheSettings);
+    return new TranslateCacheService(translateService, translateCacheSettings);
 }
 
 @NgModule({
-  declarations: [
-    AppComponent
-  ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    TranslateModule.forRoot({
-      loader: {
-        provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient]
-      }
-    }),
-    TranslateCacheModule.forRoot({
-      cacheService: {
-        provide: TranslateCacheService,
-        useFactory: TranslateCacheFactory,
-        deps: [ TranslateService, TranslateCacheSettings ]
-      }
-    })
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
+    declarations: [
+        AppComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        TranslateModule.forRoot({
+            loader: {
+                provide: TranslateLoader,
+                useFactory: HttpLoaderFactory,
+                deps: [HttpClient]
+            }
+        }),
+        TranslateCacheModule.forRoot({
+            cacheService: {
+                provide: TranslateCacheService,
+                useFactory: TranslateCacheFactory,
+                deps: [TranslateService, TranslateCacheSettings]
+            }
+        })], providers: [provideHttpClient(withInterceptorsFromDi())]
 })
 export class AppModule { }
